DB/Style updates
- Add image_path attribute to Puzzle model.
  puzzle.models.create_puzzle now saves the
  original image and a thumbnail.

  The thumbnail is loaded on the puzzle listing
  page along with the lists of sessions.

 - Use bootstrap for style.

 - Middle mouse exclusively moves the puzzle
   viewport, even if the pointer is over a puzzle
   piece.

+++ jigs/apps/puzzle/models.py
@@ -1,6 +1,7 @@
 import UUIDField
+from image_utils import create_thumbnail
 
-from PIL import Image
+from PIL import Image, ImageOps
 
 from django.db import models,transaction
 from django.db.models.signals import post_delete
@@ -12,11 +13,16 @@ import os,shutil,random
 
 class Puzzle(models.Model):
     guid = UUIDField.UUIDField(primary_key=True)
-    
-    directory = property(fget=lambda self:os.path.join(settings.MEDIA_ROOT,self.guid))
+    image_path = models.CharField(max_length=255)
+
+    directory = property(fget=lambda self: os.path.join(settings.MEDIA_ROOT, self.guid))
     
     def get_piece_paths(self):
         return [os.path.join(self.directory,"%s.jpg"%piece.pk) for piece in self.puzzlepiece_set.all()]
+
+    def get_thumbnail_path(self, width=256, height=256):
+        return os.path.join(os.path.split(self.directory)[1],
+                            "%sx%s_%s" % (width, height, os.path.split(self.image_path)[1]))
 
 
 class PuzzlePiece(models.Model):
@@ -66,35 +72,45 @@ def start_session(puzzle):
     return session
 
 
-
-def create_puzzle(image_path,piece_width=100,piece_height=100):
+def create_puzzle(image_path,piece_width=500,piece_height=500):
     img = Image.open(image_path)
-    w,h = piece_width,piece_height
-    tiles_x,tiles_y = img.size[0]/w,img.size[1]/h
-        
+    w, h = piece_width,piece_height
+    tiles_x, tiles_y = img.size[0]/w,img.size[1]/h
+
     with transaction.commit_on_success():
         puzzle = Puzzle()
         puzzle.save()
         guid = puzzle.guid
-        
+
         piece_dir = puzzle.directory
         if not os.path.exists(piece_dir):
             os.makedirs(piece_dir)
-        
-        for y in range(0,tiles_y):
-            for x in range(0,tiles_x):
-                left=x*w
-                top=y*h
-                right=x*w+w
-                bottom=y*h+h
-                region = img.crop((left,top,right,bottom))
+
+        puzzle.image_path = os.path.join(os.path.split(puzzle.directory)[1],
+                                         os.path.split(image_path)[1])
+        # Save the original
+        img.save(os.path.join(settings.MEDIA_ROOT, puzzle.image_path))
+        # Create a thumbnail
+        thumb = create_thumbnail(img)
+        thumb.save(os.path.join(settings.MEDIA_ROOT,
+                   puzzle.get_thumbnail_path()))
+
+        for y in range(0, tiles_y):
+            for x in range(0, tiles_x):
+                left= x * w
+                top = y * h
+                right = x * w + w
+                bottom = y * h + h
+                region = img.crop((left, top, right, bottom))
                 
-                piece = PuzzlePiece(puzzle=puzzle,x=left,y=top,width=w,height=h)
+                piece = PuzzlePiece(puzzle=puzzle, x=left, y=top, width=w, height=h)
                 piece.save()
-                region.save( os.path.join(piece_dir,"%s.jpg"%piece.pk))
+                region.save( os.path.join(piece_dir, "%s.jpg" % piece.pk))
+        puzzle.save()
     return puzzle
                 
 @receiver(post_delete)
 def delete_puzzle(sender, instance, **kwargs):
     if sender == Puzzle:
-        shutil.rmtree(instance.directory)
+        if os.path.exists(instance.directory):
+            shutil.rmtree(instance.directory)
